pyproject.toml
[tool.poetry]
name = "byebyepp"
version = "0.1.0"
description = "Bye byyyye PP"
authors = ["Admin <3decision@discngine.com>"]

[tool.poetry.dependencies]
python = "^3.7"
uvicorn = "^0.11.8"
fastapi = "^0.54.1"
python-multipart = "^0.0.5"
email-validator = "^1.0.5"
redis = "^3.5.2"
requests = "^2.24.0"
celery = "^4.4.7"
flower = "^0.9.5"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
tenacity = "^6.1.0"
pydantic = "^1.6.1"
emails = "^0.5.15"
raven = "^6.10.0"
gunicorn = "^20.0.4"
jinja2 = "^2.11.2"
psycopg2-binary = "^2.8.5"
pytest = "^5.4.1"
python-jose = {extras = ["cryptography"], version = "^3.1.0"}
biopython = "^1.77"

[tool.poetry.dev-dependencies]
mypy = "^0.782"
black = "^19.10b0"
isort = "^4.3.21"
autoflake = "^1.3.1"
flake8 = "^3.7.9"
pytest = "^5.4.1"
pytest-cov = "^2.8.1"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
